+++ Assets/Nathan/PositionHolder.cs
@@ -7,8 +7,8 @@ public class PositionHolder
     public class PositionData
     {
         public bool isoffense; //offense = true, defense = false
-        public int x;
-        public int y;
+        public float x;
+        public float y;
         public bool hasPuck;
 
     }
@@ -16,13 +16,13 @@ public class PositionHolder
 
     public class PuckData
     {
-        public int puckx;
-        public int pucky;
+        public float puckx;
+        public float pucky;
     }
 
     public PuckData puck = new PuckData();
 
-    public void updatepuck(int newpuckx, int newpucky)
+    public void updatepuck(float newpuckx, float newpucky)
     {
         puck.puckx = newpuckx;
         puck.pucky = newpucky;
@@ -31,7 +31,7 @@ public class PositionHolder
 
     private Dictionary<string, PositionData> positions = new Dictionary<string, PositionData>();
 
-    public void update(string key, bool isoffense, int x, int y, bool hasPuck)
+    public void update(string key, bool isoffense, float x, float y, bool hasPuck)
     {
         positions[key] = new PositionData { isoffense = isoffense, x = x, y = y, hasPuck = hasPuck };
     }
